Backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pymongo==4.10.1
python-multipart
python-dotenv